+++ app/src/main/java/com/example/ub_eats/Menu/myAdapterr.java
@@ -1,24 +1,24 @@
-package com.example.ub_eats;
+package com.example.ub_eats.Menu;
 
 import android.content.Context;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
-public class MyAdapter extends RecyclerView.Adapter<MyAdapter.MyViewHolder> {
+import com.example.ub_eats.R;
 
+public class myAdapterr extends RecyclerView.Adapter<myAdapterr.MyViewHolderr> {
     String data1[], data2[];
     int images[];
     Context context;
 
-
-    public MyAdapter(Context ct, String s1[], String s2[], int img[]){
-
+    public myAdapterr(Context ct, String s1[], String s2[], int img[]){
         context=ct;
         data1=s1;
         data2=s2;
@@ -27,14 +27,16 @@ public class MyAdapter extends RecyclerView.Adapter<MyAdapter.MyViewHolder> {
 
     @NonNull
     @Override
-    public MyViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
+    public MyViewHolderr onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
         LayoutInflater inflater= LayoutInflater.from(context);
-        View view=inflater.inflate(R.layout.my_row, parent,false);
-        return new MyViewHolder(view);
+        View view=inflater.inflate(R.layout.my_row2, parent,false);
+        return new MyViewHolderr(view);
+
+                //return null;
     }
 
     @Override
-    public void onBindViewHolder(@NonNull MyViewHolder holder, int position) {
+    public void onBindViewHolder(@NonNull MyViewHolderr holder, int position) {
 
         holder.mytext1.setText(data1[position]);
         holder.mytext2.setText(data2[position]);
@@ -44,21 +46,23 @@ public class MyAdapter extends RecyclerView.Adapter<MyAdapter.MyViewHolder> {
 
     @Override
     public int getItemCount() {
-
         return images.length;
     }
 
-    public class MyViewHolder extends RecyclerView.ViewHolder{
+    public class MyViewHolderr extends RecyclerView.ViewHolder {
+
 
         TextView mytext1, mytext2;
         ImageView myImage;
+        Button bt;
 
-        public MyViewHolder(@NonNull View itemView) {
-
+        public MyViewHolderr(@NonNull View itemView) {
             super(itemView);
+
             mytext1=itemView.findViewById(R.id.myText1);
             mytext2=itemView.findViewById(R.id.myText2);
             myImage=itemView.findViewById(R.id.myImageView);
+            bt=itemView.findViewById(R.id.button);
         }
     }
 }
