+++ app/src/main/res/layout/activity_main2.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Main2Activity">
+    tools:context=".Menu.Main2Activity">
 
     <TextView
         android:id="@+id/textView3"
