+++ app/src/main/res/layout/activity_second.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".secondActivity">
+    tools:context=".Menu.secondActivity">
 
     <ImageView
         android:id="@+id/imageView"
