+++ app/src/main/res/layout/fragment_item_list.xml
@@ -9,5 +9,5 @@
     android:layout_marginLeft="16dp"
     android:layout_marginRight="16dp"
     app:layoutManager="LinearLayoutManager"
-    tools:context=".ItemFragment"
+    tools:context=".Menu.ItemFragment"
     tools:listitem="@layout/fragment_item" />
